Calendario individual, arreglado para que no se muestren rutas si no eres Admin.

+++ app/Http/Controllers/FullCalendarEventMasterController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Event;
+use Auth;
+use Error;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Redirect, Response;
@@ -18,7 +20,7 @@ class FullCalendarEventMasterController extends Controller
      */
     public function __construct()
     {
-        $this->middleware(['admin']);
+        $this->middleware(['auth']);
     }
 
     /**
@@ -37,7 +39,10 @@ class FullCalendarEventMasterController extends Controller
     public function index()
     {
         if (request()->ajax()) {
-            $data = DB::table('events')->get();
+            if(Auth::user()->admin == 1)
+                $data = DB::table('events')->get();
+            else
+                $data = DB::table('events')->get()->where('name_user', '=', Auth::user()->name);
             return Response::json($data);
         }
         return view('pages.calendar');
@@ -45,42 +50,60 @@ class FullCalendarEventMasterController extends Controller
 
     public function create(Request $request)
     {
-        $this->validator($request->all())->validate();
+        if(Auth::user()->admin == 1)
+        {
+            $this->validator($request->all())->validate();
 
-        $insertArr = [
-            'title' => $request->title,
-            'start' => $request->start,
-            'end' => $request->end,
-            'description'=>$request->description,
-            'name_user'=>$request->name_user,
-            'id_user'=>$request->id_user
-        ];
-        $event = Event::create($insertArr)->id;
-        return Response::json($event);
+            $insertArr = [
+                'title' => $request->title,
+                'start' => $request->start,
+                'end' => $request->end,
+                'description'=>$request->description,
+                'name_user'=>$request->name_user,
+                'id_user'=>$request->id_user
+            ];
+            $event = Event::create($insertArr)->id;
+            return Response::json($event);
+        }else
+            return abort(403, 'Esta acción no está permitida');        
     }
 
     public function update(Request $request)
     {
-        $where = array('id' => $request->id);
-        $updateArr = ['title' => $request->title, 'start' => $request->start, 'end' => $request->end];
-        $event  = Event::where($where)->update($updateArr);
-        return Response::json($event);
+        if(Auth::user()->admin == 1)
+        {
+            $where = array('id' => $request->id);
+            $updateArr = ['title' => $request->title, 'start' => $request->start, 'end' => $request->end];
+            $event  = Event::where($where)->update($updateArr);
+            return Response::json($event);
+        }else
+            return abort(403, 'Esta acción no está permitida');        
     }
 
     public function edit(Request $request)
     {
-        $where = array('id' => $request->id);
-        $editArr = ['title' => $request->title, 'start' => $request->start, 'end' => $request->end, 
-                    'name_user' => $request->name_user, 'id_user' => $request->id_user, 
-                    'description' => $request->description];
-        $event = Event::where($where)->update($editArr);
-        return Response::json($event);
+        if(Auth::user()->admin == 1)
+        {
+            $where = array('id' => $request->id);
+            $editArr = ['title' => $request->title, 'start' => $request->start, 'end' => $request->end, 
+                        'name_user' => $request->name_user, 'id_user' => $request->id_user, 
+                        'description' => $request->description];
+            $event = Event::where($where)->update($editArr);
+            return Response::json($event);
+        }else
+            return abort(403, 'Esta acción no está permitida');
+       
     }
 
     public function destroy(Request $request)
     {
-        $event = Event::where('id', $request->id)->delete();
-        return Response::json($event);
+        if(Auth::user()->admin == 1)
+        {
+            $event = Event::where('id', $request->id)->delete();
+            return Response::json($event);
+        }else 
+            return abort(403, 'Esta acción no está permitida');
+        
     }
 
 }
